+++ model/src/main/java/com/idiot/model/RecommendedEstate.kt
@@ -3,7 +3,7 @@ package com.idiot.model
 import com.squareup.moshi.Json
 
 data class RecommendedEstate(
-    @field:Json(name = "id") val id: Long,
+    @field:Json(name = "id") val id: Int,
     @field:Json(name = "estateThumbNail") val estateThumbNail: String,
     @field:Json(name = "town") val town: String,
     @field:Json(name = "thumbnail3D") val thumbnail3D: String,
